--- ecosystem.config.js
@@ -1,13 +1,0 @@
-module.exports = {
-  apps : [
-    {
-      name      : 'better-place',
-      script    : 'bin/www',
-      env: {
-      },
-      env_production : {
-        NODE_ENV: 'production',
-      },
-    },
-  ],
-};
