+++ package.json
@@ -6,7 +6,7 @@
   "homepage": "https://github.com/michelbl/better-place#readme",
   "license": "ISC",
   "scripts": {
-    "start": "node ./bin/www",
+    "start": "node ./bin/run.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
